requirements.txt
alembic==1.7.5
blinker==1.7.0
click==8.0.3
Faker==24.4.0
Flask==1.1.4
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.2.0
six==1.16.0
SQLAlchemy==1.4.31
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==2.0.2
wheel==0.43.0
